+++ packages/dart_edge_http_server_runtime/lib/src/runtime/dart_edge_http_server.dart
@@ -4,6 +4,8 @@ import 'dart:ffi';
 import 'dart:io';
 
 import 'package:dart_edge_core/dart_edge_core.dart';
+import 'package:dart_edge_native_bridge/dart_edge_native_bridge.dart'
+    show NativeBinaryPayloadLease;
 import 'package:json_schema/json_schema.dart';
 
 import '../native/dart_edge_native.dart';