+++ packages/dart_edge_http_server_runtime/test/dart_edge_runtime_http_test.dart
@@ -4,6 +4,8 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:dart_edge_http_server_runtime/dart_edge_http_server_runtime.dart';
+import 'package:dart_edge_native_bridge/dart_edge_native_bridge.dart'
+    show NativeBinaryPayloadLease;
 import 'package:test/test.dart';
 
 void main() {