+++ packages/dart_edge_vad/lib/src/native_vad_stream.dart
@@ -2,7 +2,8 @@ import 'dart:convert';
 import 'dart:ffi';
 import 'dart:typed_data';
 
-import 'package:dart_edge_core/dart_edge_core.dart';
+import 'package:dart_edge_native_bridge/dart_edge_native_bridge.dart'
+    show NativeBinaryPayloadLease;
 
 import 'native/dart_edge_vad_native.dart';
 import 'native/generated_bindings.dart' as gen;