+++ packages/dart_edge_vad/test/dart_edge_vad_test.dart
@@ -1,7 +1,8 @@
 import 'dart:ffi';
 import 'dart:typed_data';
 
-import 'package:dart_edge_core/dart_edge_core.dart';
+import 'package:dart_edge_native_bridge/dart_edge_native_bridge.dart'
+    show NativeBinaryPayloadLease;
 import 'package:dart_edge_vad/dart_edge_vad.dart';
 import 'package:dart_edge_vad/src/native/dart_edge_vad_native.dart';
 import 'package:ffi/ffi.dart';